Added inline version of README example workflow.

+++ src/main/resources/org/kurator/akka/readme/reader.py
@@ -2,5 +2,5 @@ import sys
 
 def readnum():
     line = sys.stdin.readline()
-    n = int(line.strip())
+    n = float(line.strip())
     return n
